Sorted commands in to folders based on the subsystem they belong to, fixed imports
Commands that interact with multiple subsystems are sorted
