+++ InfiniteRecharge2020-PBot/src/main/java/frc/robot/commands/CombinationCommandGroups/BallLoadingCommandGroup.java
@@ -5,15 +5,18 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.CombinationCommandGroups;
 
-import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import frc.robot.subsystems.ExampleSubsystem;
+import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
+import frc.robot.commands.HopperCommands.MoveBall;
+import frc.robot.commands.IntakeCommands.IntakeOn;
+import frc.robot.subsystems.HopperSubsystem;
+import frc.robot.subsystems.IntakeSubsystem;
 
 /**
  * A complex auto command that drives forward, releases a hatch, and then drives backward.
  */
-public class ExampleCommandGroup extends SequentialCommandGroup {
+public class BallLoadingCommandGroup extends ParallelCommandGroup {
   /**
    * Creates a new Command Group.
    * There are 4 types of command groups:
@@ -25,13 +28,13 @@ public class ExampleCommandGroup extends SequentialCommandGroup {
    * @param subsystem1 The subsystem this command will run on
    * @param subsystem2 The subsystem this command will run on
    */
-  public ExampleCommandGroup(ExampleSubsystem subsystem1, ExampleSubsystem subsystem2) {
+  public BallLoadingCommandGroup(HopperSubsystem hopperSubsystem, IntakeSubsystem intakeSubsystem) {
     addCommands(
-        // Example 1
-        new ExampleCommand(subsystem1),
+        // Turns on intake
+        new IntakeOn(intakeSubsystem),
 
-        // Example 2
-        new ExampleCommand(subsystem2)
+        // Does move ball command 
+        new MoveBall(hopperSubsystem, true, true)
     );
   }
