+++ InfiniteRecharge2020-PBot/src/main/java/frc/robot/commands/CombinationCommandGroups/BallLoadingCommandGroupStop.java
@@ -5,41 +5,36 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands.CommandGroups;
+package frc.robot.commands.CombinationCommandGroups;
 
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
-import frc.robot.commands.HopperCommands.TurnOnAllHopperMotors;
-import frc.robot.commands.IntakeCommands.IntakeOn;
-import frc.robot.commands.ShooterCommands.ShooterOn;
+import frc.robot.commands.HopperCommands.HopperOff;
+import frc.robot.commands.IntakeCommands.IntakeOff;
 import frc.robot.subsystems.HopperSubsystem;
 import frc.robot.subsystems.IntakeSubsystem;
-import frc.robot.subsystems.ShooterSubsystem;
 
 /**
  * A complex auto command that drives forward, releases a hatch, and then drives backward.
  */
-public class TurnOnAllMotors extends ParallelCommandGroup {
+public class BallLoadingCommandGroupStop extends ParallelCommandGroup {
   /**
    * Creates a new Command Group.
    * There are 4 types of command groups:
    * SequentialCommandGroup - first command is executed, then the second, etc. and so on until the list finishes
    * ParallelCommandGroup - ends commands when they finish
    * ParallelRaceGroup - ends as soon as any command in the group ends
-   * ParallelDeadlineGroup - ends when a specific command (the "deadline") ends
+   * ParallelDeadlineGroup - ends when a specific command (the “deadline”) ends
    *
    * @param subsystem1 The subsystem this command will run on
    * @param subsystem2 The subsystem this command will run on
    */
-  public TurnOnAllMotors(HopperSubsystem hopperSubsystem, IntakeSubsystem intakeSubsystem, ShooterSubsystem shooterSubsystem) {
+  public BallLoadingCommandGroupStop(HopperSubsystem hopper, IntakeSubsystem intake) {
     addCommands(
-        // Turns on hopper
-        new TurnOnAllHopperMotors(hopperSubsystem),
+        // Turns off intake
+        new IntakeOff(intake),
 
-        // Turns on intake
-        new IntakeOn(intakeSubsystem),
-
-        // Turns on shooter
-        new ShooterOn(shooterSubsystem)
+        // Turns off hopper motor
+        new HopperOff(hopper)
     );
   }
