+++ InfiniteRecharge2020-PBot/src/main/java/frc/robot/commands/CombinationCommandGroups/IntakeandSuckerOff.java
@@ -5,35 +5,36 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands.CommandGroups;
+package frc.robot.commands.CombinationCommandGroups;
 
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
-import frc.robot.commands.HopperCommands.ReverseAllHopperMotors;
-import frc.robot.commands.IntakeCommands.IntakeReverse;
+import frc.robot.commands.HopperCommands.SuckerOff;
+import frc.robot.commands.IntakeCommands.IntakeOff;
 import frc.robot.subsystems.HopperSubsystem;
 import frc.robot.subsystems.IntakeSubsystem;
 
 /**
  * A complex auto command that drives forward, releases a hatch, and then drives backward.
  */
-public class ReverseAllMotorsExceptShooter extends ParallelCommandGroup {
+public class IntakeandSuckerOff extends ParallelCommandGroup {
   /**
-   * Creates a new Command Group. There are 4 types of command groups:
-   * SequentialCommandGroup - first command is executed, then the second, etc. and
-   * so on until the list finishes ParallelCommandGroup - ends commands when they
-   * finish ParallelRaceGroup - ends as soon as any command in the group ends
-   * ParallelDeadlineGroup - ends when a specific command (the "deadline") ends
+   * Creates a new Command Group.
+   * There are 4 types of command groups:
+   * SequentialCommandGroup - first command is executed, then the second, etc. and so on until the list finishes
+   * ParallelCommandGroup - ends commands when they finish
+   * ParallelRaceGroup - ends as soon as any command in the group ends
+   * ParallelDeadlineGroup - ends when a specific command (the “deadline”) ends
    *
    * @param subsystem1 The subsystem this command will run on
    * @param subsystem2 The subsystem this command will run on
    */
-  public ReverseAllMotorsExceptShooter(HopperSubsystem hopperSubsystem, IntakeSubsystem intakeSubsystem) {
+  public IntakeandSuckerOff(IntakeSubsystem intakeSubsystem, HopperSubsystem hopperSubsystem) {
     addCommands(
-        // Reverses hopper
-        new ReverseAllHopperMotors(hopperSubsystem),
+        // Turns off intake
+        new IntakeOff(intakeSubsystem),
 
-        // Reverses intake
-        new IntakeReverse(intakeSubsystem)
+        // Turns off sucker
+        new SuckerOff(hopperSubsystem)
     );
   }
