+++ InfiniteRecharge2020-PBot/src/main/java/frc/robot/commands/CombinationCommandGroups/IntakeandSuckerOn.java
@@ -5,41 +5,36 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands.CommandGroups;
+package frc.robot.commands.CombinationCommandGroups;
 
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
-import frc.robot.commands.HopperCommands.TurnOffAllHopperMotors;
-import frc.robot.commands.IntakeCommands.IntakeOff;
-import frc.robot.commands.ShooterCommands.ShooterOff;
+import frc.robot.commands.HopperCommands.SuckerOn;
+import frc.robot.commands.IntakeCommands.IntakeOn;
 import frc.robot.subsystems.HopperSubsystem;
 import frc.robot.subsystems.IntakeSubsystem;
-import frc.robot.subsystems.ShooterSubsystem;
 
 /**
  * A complex auto command that drives forward, releases a hatch, and then drives backward.
  */
-public class TurnOffAllMotors extends ParallelCommandGroup {
+public class IntakeandSuckerOn extends ParallelCommandGroup {
   /**
    * Creates a new Command Group.
    * There are 4 types of command groups:
    * SequentialCommandGroup - first command is executed, then the second, etc. and so on until the list finishes
    * ParallelCommandGroup - ends commands when they finish
    * ParallelRaceGroup - ends as soon as any command in the group ends
-   * ParallelDeadlineGroup - ends when a specific command (the "deadline") ends
+   * ParallelDeadlineGroup - ends when a specific command (the “deadline”) ends
    *
    * @param subsystem1 The subsystem this command will run on
    * @param subsystem2 The subsystem this command will run on
    */
-  public TurnOffAllMotors(HopperSubsystem hopperSubsystem, IntakeSubsystem intakeSubsystem, ShooterSubsystem shooterSubsystem ) {
+  public IntakeandSuckerOn(IntakeSubsystem intakeSubsystem, HopperSubsystem hopperSubsystem) {
     addCommands(
-        // Turns off hopper
-        new TurnOffAllHopperMotors(hopperSubsystem),
-        
-        // Turns off intake
-        new IntakeOff(intakeSubsystem),
+        // Turns on intake
+        new IntakeOn(intakeSubsystem),
 
-        // Turns off shooter
-        new ShooterOff(shooterSubsystem)
+        // Turns on sucker
+        new SuckerOn(hopperSubsystem)
     );
   }
