+++ InfiniteRecharge2020-PBot/src/main/java/frc/robot/commands/CombinationCommandGroups/ReverseAllMotorsExceptShooter.java
@@ -5,36 +5,35 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands.CommandGroups;
+package frc.robot.commands.CombinationCommandGroups;
 
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
-import frc.robot.commands.HopperCommands.SuckerOn;
-import frc.robot.commands.IntakeCommands.IntakeOn;
+import frc.robot.commands.HopperCommands.ReverseAllHopperMotors;
+import frc.robot.commands.IntakeCommands.IntakeReverse;
 import frc.robot.subsystems.HopperSubsystem;
 import frc.robot.subsystems.IntakeSubsystem;
 
 /**
  * A complex auto command that drives forward, releases a hatch, and then drives backward.
  */
-public class IntakeandSuckerOn extends ParallelCommandGroup {
+public class ReverseAllMotorsExceptShooter extends ParallelCommandGroup {
   /**
-   * Creates a new Command Group.
-   * There are 4 types of command groups:
-   * SequentialCommandGroup - first command is executed, then the second, etc. and so on until the list finishes
-   * ParallelCommandGroup - ends commands when they finish
-   * ParallelRaceGroup - ends as soon as any command in the group ends
-   * ParallelDeadlineGroup - ends when a specific command (the “deadline”) ends
+   * Creates a new Command Group. There are 4 types of command groups:
+   * SequentialCommandGroup - first command is executed, then the second, etc. and
+   * so on until the list finishes ParallelCommandGroup - ends commands when they
+   * finish ParallelRaceGroup - ends as soon as any command in the group ends
+   * ParallelDeadlineGroup - ends when a specific command (the "deadline") ends
    *
    * @param subsystem1 The subsystem this command will run on
    * @param subsystem2 The subsystem this command will run on
    */
-  public IntakeandSuckerOn(IntakeSubsystem intakeSubsystem, HopperSubsystem hopperSubsystem) {
+  public ReverseAllMotorsExceptShooter(HopperSubsystem hopperSubsystem, IntakeSubsystem intakeSubsystem) {
     addCommands(
-        // Turns on intake
-        new IntakeOn(intakeSubsystem),
+        // Reverses hopper
+        new ReverseAllHopperMotors(hopperSubsystem),
 
-        // Turns on sucker
-        new SuckerOn(hopperSubsystem)
+        // Reverses intake
+        new IntakeReverse(intakeSubsystem)
     );
   }
