+++ InfiniteRecharge2020-PBot/src/main/java/frc/robot/commands/ExampleCommands/ExampleCommand.java
@@ -5,46 +5,38 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.ExampleCommands;
 
-import frc.robot.subsystems.HopperSubsystem;
-import frc.robot.subsystems.IntakeSubsystem;
+import frc.robot.subsystems.ExampleSubsystem;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 
 /**
  * An example command that uses an example subsystem.
  */
-public class IntakeOneExtraBall extends CommandBase {
+public class ExampleCommand extends CommandBase {
   @SuppressWarnings({"PMD.UnusedPrivateField", "PMD.SingularField"})
-  private final HopperSubsystem m_hopper;
-  private IntakeSubsystem m_intake;
+  private final ExampleSubsystem m_subsystem;
 
   /**
    * Creates a new ExampleCommand.
    *
    * @param subsystem The subsystem used by this command.
    */
-  public IntakeOneExtraBall(HopperSubsystem hopper, IntakeSubsystem intake) {
-    m_hopper = hopper;
-    m_intake = intake;
-
+  public ExampleCommand(ExampleSubsystem subsystem) {
+    m_subsystem = subsystem;
     // Use addRequirements() here to declare subsystem dependencies.
-    addRequirements(hopper);
-    addRequirements(intake);
+    addRequirements(subsystem);
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-    System.out.println("Intaking one extra ball");
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    m_intake.intakeOn();   // Make sure intake is on
-    m_hopper.setIntakeMode();  // Make sure we are in intake mode
-    m_hopper.intakeOneBallNoCountCheck();  // Now feed the ball through the hopper while it is still sensed
+    m_subsystem.dummy();
   }
 
   // Called once the command ends or is interrupted.
