+++ InfiniteRecharge2020-PBot/src/main/java/frc/robot/commands/SetHopperModeCommands/SetIntakeMode.java
@@ -5,18 +5,20 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.SetHopperModeCommands;
 
 import frc.robot.subsystems.HopperSubsystem;
+import frc.robot.subsystems.IntakeSubsystem;
 import frc.robot.subsystems.ShooterSubsystem;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 
 /**
  * An example command that uses an example subsystem.
  */
-public class ShootAllBalls extends CommandBase {
+public class SetIntakeMode extends CommandBase {
   @SuppressWarnings({"PMD.UnusedPrivateField", "PMD.SingularField"})
   private final HopperSubsystem m_hopper;
+  private final IntakeSubsystem m_intake;
   private final ShooterSubsystem m_shooter;
 
   /**
@@ -24,29 +26,30 @@ public class ShootAllBalls extends CommandBase {
    *
    * @param subsystem The subsystem used by this command.
    */
-  public ShootAllBalls(HopperSubsystem hopper, ShooterSubsystem shooter) {
+  public SetIntakeMode(HopperSubsystem hopper, IntakeSubsystem intake, ShooterSubsystem shooter) {
     m_hopper = hopper;
+    m_intake = intake;
     m_shooter = shooter;
 
     // Use addRequirements() here to declare subsystem dependencies.
     addRequirements(hopper);
+    addRequirements(intake);
     addRequirements(shooter);
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-    System.out.println("Shooting one ball");
+    System.out.println("Setting Intake Mode");
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    m_shooter.shooterOn();
-    m_hopper.setShootingMode();  // Make sure we are in shooting mode
-    m_hopper.outputAllBalls();  // Now feed all of the ball through the hopper 
-    // Note that this command does not end. Assumption is that once balls stop coming out, operator
-    // will change modes to intake mode which will turn off the shooter
+    m_hopper.setIntakeMode();
+    m_hopper.intakeOneBall();
+    m_intake.intakeOn();
+    m_shooter.shooterOff();
   }
 
   // Called once the command ends or is interrupted.
