+++ InfiniteRecharge2020-PBot/src/main/java/frc/robot/commands/SetHopperModeCommands/SetShootingMode.java
@@ -5,7 +5,7 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands.RobotModeCommands;
+package frc.robot.commands.SetHopperModeCommands;
 
 import frc.robot.subsystems.HopperSubsystem;
 import frc.robot.subsystems.IntakeSubsystem;
@@ -15,40 +15,40 @@ import edu.wpi.first.wpilibj2.command.CommandBase;
 /**
  * An example command that uses an example subsystem.
  */
-public class SetOffMode extends CommandBase {
+public class SetShootingMode extends CommandBase {
   @SuppressWarnings({"PMD.UnusedPrivateField", "PMD.SingularField"})
   private final HopperSubsystem m_hopper;
-  private final IntakeSubsystem m_intake;
   private final ShooterSubsystem m_shooter;
+  private IntakeSubsystem m_intake;
 
   /**
    * Creates a new ExampleCommand.
    *
    * @param subsystem The subsystem used by this command.
    */
-  public SetOffMode(HopperSubsystem hopper, IntakeSubsystem intake, ShooterSubsystem shooter) {
+  public SetShootingMode(HopperSubsystem hopper, ShooterSubsystem shooter, IntakeSubsystem intake) {
     m_hopper = hopper;
-    m_intake = intake;
     m_shooter = shooter;
+    m_intake = intake;
     
     // Use addRequirements() here to declare subsystem dependencies.
     addRequirements(hopper);
-    addRequirements(intake);
     addRequirements(shooter);
+    addRequirements(intake);
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-    System.out.println("Setting Hopper Off Mode");
+    System.out.println("Setting Shooter Mode");
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    m_hopper.setOffMode();
+    m_hopper.setShootingMode();
+    m_shooter.shooterOn();
     m_intake.intakeOff();
-    m_shooter.shooterOff();
   }
 
   // Called once the command ends or is interrupted.
@@ -59,6 +59,6 @@ public class SetOffMode extends CommandBase {
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return true;
+    return false;
   }
 }
