+++ InfiniteRecharge2020-PBot/src/main/java/frc/robot/commands/UseHopperModeCommands/ShootOneBall.java
@@ -5,38 +5,46 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.UseHopperModeCommands;
 
-import frc.robot.subsystems.ExampleSubsystem;
+import frc.robot.subsystems.HopperSubsystem;
+import frc.robot.subsystems.ShooterSubsystem;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 
 /**
  * An example command that uses an example subsystem.
  */
-public class ExampleCommand extends CommandBase {
+public class ShootOneBall extends CommandBase {
   @SuppressWarnings({"PMD.UnusedPrivateField", "PMD.SingularField"})
-  private final ExampleSubsystem m_subsystem;
+  private final HopperSubsystem m_hopper;
+  private ShooterSubsystem m_shooter;
 
   /**
    * Creates a new ExampleCommand.
    *
    * @param subsystem The subsystem used by this command.
    */
-  public ExampleCommand(ExampleSubsystem subsystem) {
-    m_subsystem = subsystem;
+  public ShootOneBall(HopperSubsystem hopper, ShooterSubsystem shooter) {
+    m_hopper = hopper;
+    m_shooter = shooter;
+
     // Use addRequirements() here to declare subsystem dependencies.
-    addRequirements(subsystem);
+    addRequirements(hopper);
+    addRequirements(shooter);
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
+    System.out.println("Shooting one ball");
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    m_subsystem.dummy();
+    m_shooter.shooterOn();
+    m_hopper.setShootingMode();  // Make sure we are in shooting mode
+    m_hopper.outputOneBall();  // Now feed the ball through the hopper while it is still sensed
   }
 
   // Called once the command ends or is interrupted.
